requirements.txt
numpy
pandas
torch
torchvision
matplotlib
Pillow
tqdm

